server/requirements.txt
Flask==1.0.2 
Flask-Cors==3.0.7
absl-py==0.10.0
appdirs==1.4.4
astroid==2.4.2
astunparse==1.6.3
black==20.8b1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
gast==0.3.3
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
isort==5.4.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
matplotlib==3.3.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
pandas==1.1.1
pathspec==0.8.0
Pillow==7.2.0
progressbar2==3.52.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wandb==0.10.8
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
joblib==0.17.0
scikit-learn